Extract getDomainScale & getRangeScale as util fns
These functions are nearly identical in Bars, GroupedBars, and StackedBars, so observe DRY principle.
